ForGOOS function
This is a dynamic way to select a path implementation based on a GOOS
string, using the same mappings as we use at compile time to populate
"Target", for programs that are e.g. taking part in a Go build process
themselves and want to generate something for the target, or that just
find it more convenient to work with runtime.GOOS for other reasons.